+++ R/FeatureImp.R
@@ -9,9 +9,9 @@
 #' 
 #' @section Usage:
 #' \preformatted{
-#' imp = FeatureImp$new(predictor, loss, method = "shuffle", compare = "ratio", n.repetitions = 3, run = TRUE)
+#' imp = FeatureImp$new(predictor, loss, compare = "ratio", n.repetitions = 5, run = TRUE)
 #' 
-#' plot(imp)
+#' imp$plot()
 #' imp$results
 #' print(imp)
 #' } 
@@ -22,26 +22,27 @@
 #' \describe{
 #' \item{predictor: }{(Predictor)\cr 
 #' The object (created with Predictor$new()) holding the machine learning model and the data.}
-#' \item{loss: }{(`character(1)` | function)\cr The loss function. Either the name of a loss (e.g. "ce" for classification or "mse") or a loss function. See Details for allowed losses.}
-#' \item{method: }{(`character(1)`\cr Either "shuffle" or "cartesian". See Details.}
+#' \item{loss: }{(`character(1)` | function)\cr The loss function. Either the name of a loss (e.g. "ce" for classification or "mse") or a function. See Details for allowed losses.}
 #' \item{compare: }{(`character(1)`)\cr Either "ratio" or "difference". 
 #' Should importance be measured as the difference or as the ratio of original model error and model error after permutation?
 #' Ratio: error.permutation/error.orig,   Difference: error.permutation - error.orig}
-#' \item{n.repetitions: }{`numeric(1)`\cr How often should the shuffling of the feature be repeated? Ignored if method is set to "cartesian".
-#' The higher the number of repetitions the more stable the results will become.}
+#' \item{n.repetitions: }{`numeric(1)`\cr How often should the shuffling of the feature be repeated?
+#' The higher the number of repetitions the more stable and accurate the results become. 
+#' }
 #' \item{parallel: }{`logical(1)`\cr Should the method be executed in parallel? If TRUE, requires a cluster to be registered, see ?foreach::foreach.}
 #' \item{run: }{(`logical(1)`)\cr Should the Interpretation method be run?}
 #' }
 #' 
 #' @section Details:
-#' Read the Interpretable Machine Learning book to learn in detail about feature importance: 
+#' 
+#' To compute the feature importance for a single feature, the model prediction loss (error) is measured before
+#' and after shuffling the values of the feature. By shuffling the feature values, the association between the 
+#' outcome and the feature is destroyed. The larger the increase in prediction error, the more important
+#' the feature was. The shuffling is repeated to get more accurate results, since the permutation 
+#' feature importance tends to be quite instable.
+#' Read the Interpretable Machine Learning book to learn about feature importance  in detail: 
 #' \url{https://christophm.github.io/interpretable-ml-book/feature-importance.html}
 #' 
-#' Two permutation schemes are implemented: 
-#' \itemize{
-#' \item shuffle: A simple shuffling of the feature values, yielding n perturbed instances per feature (fast)
-#' \item cartesian: Matching every instance with the feature value of all other instances, yielding n x (n-1) perturbed instances per feature (very slow)
-#' }
 #' 
 #' The loss function can be either specified via a string, or by handing a function to \code{FeatureImp()}.
 #' If you want to use your own loss function it should have this signature: function(actual, predicted).
@@ -55,8 +56,12 @@
 #' \describe{
 #' \item{original.error: }{(`numeric(1)`)\cr The loss of the model before perturbing features.}
 #' \item{predictor: }{(Predictor)\cr The prediction model that was analysed.}
-#' #' \item{compare: }{(`character(1)`)\cr Either "ratio" or "difference", depending on whether the importance was calculated as difference between original model error and model error after permutation or as ratio.}
-#' \item{results: }{(data.frame)\cr data.frame with the results of the feature importance computation.}
+#' \item{compare: }{(`character(1)`)\cr Either "ratio" or "difference", depending on whether the importance was calculated as difference between original model error and model error after permutation or as ratio.}
+#' \item{results: }{(data.frame)\cr data.frame with the results of the feature importance computation. One row per feature with the following columns:
+#'                  importance.05 (5% quantile of importance values from the repetitions), 
+#'                  importance (median importance),
+#'                  importance.95 (95% quantile) and the permutation.error (median error over all repetitions). The distribution of the importance 
+#'                  is also visualized as a bar in the plots, the median importance over the repetitions as a point.}
 #' }
 #' 
 #' @section Methods:
@@ -136,34 +141,28 @@ FeatureImp = R6::R6Class("FeatureImp",
     original.error = NULL,
     n.repetitions = NULL,
     compare = NULL,
-    initialize = function(predictor, loss, method = "shuffle", compare = "ratio", n.repetitions = 3, run = TRUE, parallel = FALSE) {
-      assert_choice(method, c("shuffle", "cartesian"))
+    initialize = function(predictor, loss, compare = "ratio", 
+                          n.repetitions = 5, run = TRUE, parallel = FALSE) {
       assert_choice(compare, c("ratio", "difference"))
       assert_number(n.repetitions)
       assert_logical(parallel)
       private$parallel = parallel
       self$compare = compare
-      if (n.repetitions > predictor$data$n.rows) {
-        message('Number of repetitions larger than number of unique permutations per row. 
-          Switching to method = "cartesian"')
-        method = "cartesian"
-      }
       if (!inherits(loss, "function")) {
         ## Only allow metrics from Metrics package
         allowedLosses = c("ce", "f1", "logLoss", "mae", "mse", "rmse", "mape", "mdae",
                  "msle", "percent_bias", "rae", "rmse", "rmsle", "rse", "rrse", "smape")        
         checkmate::assert_choice(loss, allowedLosses)
-        private$loss.string  = loss
+        private$loss_string  = loss
         loss = getFromNamespace(loss, "Metrics")
       } else {
-        private$loss.string = head(loss)
+        private$loss_string = head(loss)
       }
       if (is.null(predictor$data$y)) {
         stop("Please call Predictor$new() with the y target vector.")
       }
       super$initialize(predictor = predictor)
-      self$loss = private$set.loss(loss)
-      private$method = method
+      self$loss = private$set_loss(loss)
       private$getData = private$sampler$get.xy
       self$n.repetitions = n.repetitions
       actual = private$sampler$y[[1]]
@@ -178,9 +177,8 @@ FeatureImp = R6::R6Class("FeatureImp",
     }
   ),
   private = list(
-    method = NULL,
     # for printing
-    loss.string = NULL,
+    loss_string = NULL,
     q = function(pred) probs.to.labels(pred),
     combine.aggregations = function(agg, dat){
       if(is.null(agg)) { 
@@ -193,13 +191,11 @@ FeatureImp = R6::R6Class("FeatureImp",
       private$dataSample = private$getData()
       result = NULL
       
-      estimate.feature.imp = function(feature, data.sample, y, n.repetitions, cartesian, y.names, pred, loss) {
-        cartesian = ifelse(private$method == "cartesian", TRUE, FALSE)
-        
+      estimate_feature_imp = function(feature, data.sample, y, n.repetitions, y.names, pred, loss) {
         cnames = setdiff(colnames(data.sample), y.names)
         
         mg = iml:::MarginalGenerator$new(data.sample, data.sample, 
-          features = feature, n.sample.dist = n.repetitions, y = y, cartesian = cartesian, id.dist = TRUE)
+          features = feature, n.sample.dist = n.repetitions, y = y, cartesian = FALSE, id.dist = TRUE)
         qResults = data.table::data.table()
         y.vec = data.table::data.table()
         while(!mg$finished) {
@@ -209,12 +205,12 @@ FeatureImp = R6::R6Class("FeatureImp",
         }
         
         # AGGREGATE measurements
-        results = data.table::data.table(feature = feature, actual = y.vec[[1]], predicted = qResults[[1]])
-        results = results[, list("permutation.error" = loss(actual, predicted)), by = feature]
+        results = data.table::data.table(feature = feature, actual = y.vec[[1]], predicted = qResults[[1]], 
+          num_rep = rep(1:n.repetitions, each = nrow(data.sample)))
+        results = results[, list("permutation_error" = loss(actual, predicted)), by = list(feature, num_rep)]
         results
       }
       
-      loss = self$loss
       n.repetitions = self$n.repetitions
       data.sample = private$dataSample
       y = private$sampler$y
@@ -225,17 +221,20 @@ FeatureImp = R6::R6Class("FeatureImp",
       `%mypar%` = private$get.parallel.fct(private$parallel)
       result = foreach(feature = private$sampler$feature.names, .combine = rbind, .export = "self", 
         .packages = devtools::loaded_packages()$package, .inorder = FALSE) %mypar%
-        estimate.feature.imp(feature, data.sample = data.sample, y = y, cartesian = cartesian, 
+        estimate_feature_imp(feature, data.sample = data.sample, y = y,
           n.repetitions = n.repetitions, y.names = y.names, pred  = pred, loss = loss)
-      result$original.error = self$original.error
       if (self$compare == "ratio") {
-        result[, importance := permutation.error / self$original.error]
+        result[, importance_raw := permutation_error / self$original.error]
       } else {
-        result[, importance := permutation.error - self$original.error]
+        result[, importance_raw := permutation_error - self$original.error]
       }
+      result = result[, list("importance" = median(importance_raw),
+                             "permutation.error" = median(permutation_error),
+                             "importance.05" = quantile(importance_raw, probs = 0.05),
+                             "importance.95" = quantile(importance_raw, probs = 0.95)), by = list(feature)]
       result = result[order(result$importance, decreasing = TRUE),]
       # Removes the n column
-      result = result[,list(feature, original.error, permutation.error, importance)]
+      result = result[,list(feature, importance.05, importance, importance.95, permutation.error)]
       private$finished = TRUE
       self$results = data.frame(result)
     },
@@ -245,16 +244,17 @@ FeatureImp = R6::R6Class("FeatureImp",
         res$feature = factor(res$feature, levels = res$feature[order(res$importance)])
       }
       xstart = ifelse(self$compare == "ratio", 1, 0)
-      ggplot(res, aes(y = feature, x = importance)) + geom_point()+ 
-        geom_segment(aes(y = feature, yend = feature, x=xstart, xend = importance)) + 
-        scale_x_continuous("Feature Importance") + 
+      ggplot(res, aes(y = feature, x = importance)) +  
+        geom_segment(aes(y = feature, yend = feature, x=importance.05, xend = importance.95), size = 1.5, color = "darkslategrey") +
+        geom_point(size = 3)+
+        scale_x_continuous(sprintf("Feature Importance (loss: %s)", private$loss_string)) + 
         scale_y_discrete("Feature")
     }, 
-    set.loss = function(loss) {
+    set_loss = function(loss) {
       self$loss = loss
     }, 
     printParameters = function() {
-      cat("error function:", private$loss.string)
+      cat("error function:", private$loss_string)
     }
   )
 )
@@ -269,6 +269,14 @@ FeatureImp = R6::R6Class("FeatureImp",
 #' @param ... Further arguments for the objects plot function
 #' @return ggplot2 plot object
 #' @export
+#' 
+#' @details 
+#' The plot shows the importance per feature.
+#' 
+#' When n.repetitions in `FeatureImp$new` was larger than 1, then we get multiple importance estimates per feature.
+#' The importances are aggregated and the plot shows the median importance per feature (as dots) and also the 90%-quantile, which 
+#' helps to understand how much variance the computation has per feature.
+#' 
 #' @seealso 
 #' \link{FeatureImp}
 #' @examples